Add more descriptive text to notification alert
Remarks:
It appears that cloudkit isn’t as flexible as I would have liked. I’m allowed to request fields from the record that triggers a notification, but I’m not allowed to traverse reference fields to get their fields. For example, if a Level triggers a notification, I want to get the level’s name (easy) and the name of the structure it belongs to. Apparently that’s not yet supported. womp. Instead, I create a subscription individually for each Level with custom alertBody text rather than data sent from the server. Oh well!

TODO:
- NSCoding for subscriptions so that user can more easily disable/enable them
- GUI for managing subscriptions (Structure granularity)
- Geofence around chapman so no notifications are sent while on campus